+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Help/Invite.cs
@@ -20,7 +20,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Help
         [Remarks("")]
         public async Task InviteDM()
         {
-            string devInviteUrl = "[[Kaguya Dev Invite]](https://discordapp.com/api/oauth2/authorize?client_id=367403886841036820&permissions=8&scope=bot)\n";
+            string devInviteUrl = "[[Kaguya Dev Invite]](https://discord.com/api/oauth2/authorize?client_id=664032361679159309&permissions=8&scope=bot)\n";
             const string inviteUrl = "[[Invite Kaguya to your server]](https://discordapp.com/oauth2/authorize?client_id=538910393918160916&scope=bot&permissions=469101694)\n";
             const string discordUrl = "[[Kaguya Support Discord]](https://discord.gg/aumCJhr)\n";